+++ packages/metadata-builders/src/dynamic-builder.ts
@@ -107,9 +107,10 @@ const _buildCodec = (
 }
 const buildCodec = withCache(_buildCodec, scale.Self, (res) => res)
 
-export const getDynamicBuilder = (metadata: V14 | V15) => {
-  const lookupData = metadata.lookup
-  const getLookupEntryDef = getLookupFn(lookupData)
+export const getDynamicBuilder = (
+  metadata: V14 | V15,
+  getLookupEntryDef = getLookupFn(metadata.lookup),
+) => {
   let _accountId = scale.AccountId()
 
   const cache = new Map()